Rename op_builder.rs to fused.rs.
