+++ src/daft-execution/src/ops/fused.rs
